♻️ Extract logic into composable
