+++ src/components/AffectsTable/AffectsTable.vue
@@ -1,213 +1,55 @@
 <script setup lang="ts">
-import { computed, onMounted, reactive, ref, toRaw, watch } from 'vue';
+import { onMounted, toRaw, watch } from 'vue';
 
-import {
-  FlexRender,
-  getCoreRowModel,
-  getPaginationRowModel,
-  getSortedRowModel,
-  getFilteredRowModel,
-  getFacetedUniqueValues,
-  useVueTable,
-} from '@tanstack/vue-table';
-import type { SortingState, Column, ColumnFiltersState, RowData } from '@tanstack/vue-table';
+import { FlexRender } from '@tanstack/vue-table';
 import { storeToRefs } from 'pinia';
 
 import PaginationControls from '@/components/AffectsTable/PaginationControls.vue';
 
 import { useFlaw } from '@/composables/useFlaw';
-import { usePagination } from '@/composables/usePagination';
 import { useAffectsModel } from '@/composables/useAffectsModel';
+import { useAffectsTable } from '@/composables/useAffectsTable';
 
 import { useSettingsStore } from '@/stores/SettingsStore';
 import SortIcon from '@/widgets/SortIcon/SortIcon.vue';
 import DebouncedInput from '@/widgets/DebouncedInput/DebouncedInput.vue';
-import type { ZodAffectType } from '@/types';
 
-import columnDefinitions from './columnDefinitions';
 import ColumnFilter from './ColumnFilter.vue';
-import { arrIncludesWithBlanks, cvssScore } from './customFilters';
 import ColumnOptions from './ColumnOptions.vue';
-
-declare module '@tanstack/table-core' {
-
-  interface TableMeta<TData extends RowData> {
-    createData(): void;
-    deleteData(rowId: string | string[]): void;
-    fileTrackers(rows: TData | TData[]): Promise<void>;
-    filingTracker: Set<string>;
-    revert(rowId: string): void;
-    updateData<T extends keyof ZodAffectType>(rowIndex: number, columnId: T, value: ZodAffectType[T]): void;
-  }
-}
 
 const { settings } = storeToRefs(useSettingsStore());
 
+const { flaw } = useFlaw();
+
 const {
-  actions: { fileTracker, initializeAffects, markModified, markNew, markRemoved, refreshData, revertAffect },
-  state: { currentAffects, hasChanges, initialAffects, modifiedAffects, newAffects, removedAffects },
+  actions: { initializeAffects },
+  state: { hasChanges, initialAffects },
 } = useAffectsModel();
 
-const { flaw } = useFlaw();
-const columns = ref(columnDefinitions());
-const sorting = ref<SortingState>([]);
-const showAll = ref(false);
-const globalFilter = ref('');
-const columnFilters = ref<ColumnFiltersState>([]);
-const totalPages = computed(() =>
-  Math.ceil((currentAffects.value.length || 0) / settings.value.affectsPerPage),
-);
-const { changePage, currentPage, pages } = usePagination(totalPages);
-const pagination = computed(() => ({
-  pageIndex: showAll.value ? 0 : currentPage.value - 1,
-  pageSize: showAll.value ? currentAffects.value.length : settings.value.affectsPerPage,
-}));
-const table = useVueTable({
-  get data() { return currentAffects.value; },
-  get columns() { return columns.value; },
-  getRowId: row => (row?._uuid || row?.uuid) ?? '',
-  columnResizeMode: 'onChange',
-  columnResizeDirection: 'ltr',
-  initialState: {
-    columnSizing: settings.value.affectsSizing,
+const {
+  actions: {
+    changeItemsPerPage,
+    changePage,
+    deleteSelectedRows,
+    fileSelectedTrackers,
+    refreshData,
+    revertAllChanges,
+    toggleColumnVisibility,
   },
   state: {
-    get columnVisibility() { return settings.value.affectsVisibility; },
-    get pagination() { return pagination.value; },
-    get sorting() { return sorting.value; },
-    get globalFilter() { return globalFilter.value; },
-    get columnFilters() { return columnFilters.value; },
-    get columnOrder() { return settings.value.affectsColumnOrder; },
-    get columnSizing() { return settings.value.affectsSizing; },
+    columnFilters,
+    currentAffects,
+    currentPage,
+    globalFilter,
+    modifiedAffects,
+    newAffects,
+    pages,
+    removedAffects,
+    showAll,
+    table,
+    totalPages,
   },
-  meta: {
-    createData: () => {
-      const uuid = crypto.randomUUID();
-      markNew(uuid);
-      currentAffects.value = [{
-        _uuid: uuid,
-        flaw: flaw.value.uuid,
-        ps_module: '',
-        ps_component: '',
-        ps_update_stream: '',
-        embargoed: flaw.value.embargoed,
-        alerts: [],
-        trackers: [],
-        cvss_scores: [],
-        tracker: null,
-      }, ...currentAffects.value];
-    },
-    updateData: (rowIndex, columnId, value) => {
-      const currentAffect = currentAffects.value[rowIndex];
-      currentAffect[columnId] = value;
-      markModified((currentAffect._uuid || currentAffect.uuid)!);
-      refreshData();
-    },
-    deleteData: (rowId) => {
-      if (!Array.isArray(rowId)) rowId = [rowId];
-      rowId.forEach(row => markRemoved(row));
-    },
-    revert: (rowId) => {
-      revertAffect(rowId);
-      refreshData();
-    },
-    fileTrackers: async (rows) => {
-      if (!Array.isArray(rows)) rows = [rows];
-
-      // Filter out already filing trackers
-      const trackersToFile = rows.filter(row => row.uuid && !table.options.meta?.filingTracker.has(row.uuid));
-
-      trackersToFile.forEach(({ uuid }) => table.options.meta?.filingTracker.add(uuid!));
-      try {
-        for (const trackerToFile of trackersToFile) {
-          await fileTracker(trackerToFile);
-          table.options.meta?.filingTracker.delete(trackerToFile.uuid!);
-        }
-      } finally {
-        refreshData();
-      }
-    },
-    filingTracker: reactive(new Set()),
-  },
-  onSortingChange: (updaterOrValue) => {
-    sorting.value =
-      typeof updaterOrValue === 'function'
-        ? updaterOrValue(sorting.value)
-        : updaterOrValue;
-  },
-  onGlobalFilterChange: (updaterOrValue) => {
-    globalFilter.value =
-      typeof updaterOrValue === 'function'
-        ? updaterOrValue(globalFilter.value)
-        : updaterOrValue;
-  },
-  onColumnFiltersChange: (updaterOrValue) => {
-    columnFilters.value =
-      typeof updaterOrValue === 'function'
-        ? updaterOrValue(columnFilters.value)
-        : updaterOrValue;
-  },
-  onColumnOrderChange: (updaterOrValue) => {
-    settings.value.affectsColumnOrder = typeof updaterOrValue === 'function'
-      ? updaterOrValue(settings.value.affectsColumnOrder)
-      : updaterOrValue;
-  },
-  onColumnVisibilityChange: (updaterOrValue) => {
-    settings.value.affectsVisibility = typeof updaterOrValue === 'function'
-      ? updaterOrValue(settings.value.affectsVisibility)
-      : updaterOrValue;
-  },
-  onColumnSizingChange: (updaterOrValue) => {
-    settings.value.affectsSizing = typeof updaterOrValue === 'function'
-      ? updaterOrValue(settings.value.affectsSizing)
-      : updaterOrValue;
-  },
-  filterFns: {
-    arrIncludesWithBlanks,
-    cvssScore,
-  },
-  getCoreRowModel: getCoreRowModel(),
-  getPaginationRowModel: getPaginationRowModel(),
-  getSortedRowModel: getSortedRowModel(),
-  getFilteredRowModel: getFilteredRowModel(),
-  getFacetedUniqueValues: getFacetedUniqueValues(),
-});
-
-function toggleColumnVisibility(column: Column<any, any>) {
-  settings.value.affectsVisibility = {
-    ...settings.value.affectsVisibility,
-    [column.id]: !column.getIsVisible(),
-  };
-
-  // Toggle global filtering based on column visibility
-  if (column.columnDef.meta?.filter !== false) {
-    column.columnDef.enableGlobalFilter = !column.getCanGlobalFilter();
-  }
-}
-
-function changeItemsPerPage(itemsCount: number) {
-  if (Number.isNaN(itemsCount)) {
-    itemsCount = 10;
-  }
-  settings.value.affectsPerPage = Math.max(1, Math.min(100, itemsCount));
-}
-
-function deleteSelectedRows() {
-  table.options.meta?.deleteData(table.getSelectedRowModel().flatRows.map(row => row.id));
-  table.resetRowSelection();
-}
-
-function revertAllChanges() {
-  currentAffects.value.forEach(({ _uuid, uuid }) => revertAffect((_uuid || uuid)!));
-  refreshData();
-}
-
-async function fileSelectedTrackers() {
-  const affectsWithoutTracker = table.getSelectedRowModel().flatRows
-    .filter(row => !row.original.tracker)
-    .map(row => row.original);
-  await table.options.meta?.fileTrackers(affectsWithoutTracker);
-}
+} = useAffectsTable();
 
 watch(() => initialAffects.value, () => refreshData());
